feat: draw box w/rounded corners to display event
NOTE: This only works for events that that fall directly on a time line
that is drawn (because of how the position for the box lines is found).
In the future, the positions should be found based on those lines that
it falls between and its position calculated based on the width between
those lines. Currently considering calculating its percentage of the way
to the next line and then moving that far to the next line.